requirements.txt
ollama
numpy
sounddevice
rapidfuzz
youtube-search-python
python-dotenv
fuzzywuzzy
python-Levenshtein
pygetwindow
pywin32
pyrect
pyautogui
pynput
aiohttp>=3.9.0
bs4
beautifulsoup4
psutil
win10toast
playsound==1.2.2
pycaw
websockets
RealtimeSTT
chromadb
sentence-transformers
types-scipy